Latest Changes on Payment controller

+++ WinkNaturals/Models/GetPaymentModel.cs
@@ -19,13 +19,14 @@ namespace WinkNaturals.Models
             public string FirstName { get; set; }
             public string LastName { get; set; }
             public string ZipCode { get; set; }
-            //public string AuthProfileId { get; set; }
-            //public string AuthCardId { get; set; }
+            
             public string Address1 { get; set; }
             public string  City { get; set; }
             public string State { get; set; }
             public string Country { get; set; }
-            public int Price { get; set; }
+
+            public decimal Price { get; set; }
+
         }
     }
 }
